requirements.txt
numpy>=1.19.1
scikit-learn>=0.23.1
scipy>=1.5.2
tensorboard>=2.3.0
pytorch>=1.5.0
torchvision>=0.6.0
torch-geometric>=1.6
pyyaml>=5.3.1
jupyter==1.0.0
ignite>=0.4.1
pillow>=7.2.0
h5py>=2.10.0
